+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+-   `nylasAccountId` and `nylasSchedulingPageId` to `FormSubmissionEventConfigration`
+-   `CreateSchedulingSubmissionEvent()` to `FormSubmissionEvent`
+
 ## 2.2.6 (2021-06-30)
 
 ### Added